+++ src/Utility/BackupManager.php
@@ -148,9 +148,12 @@ class BackupManager
      * @return array{headers: string, message: string}
      * @throws \LogicException
      * @since 1.1.0
+     * @deprecated 2.13.4: `send()` method is deprecated. Will be removed in a future release
      */
     public static function send(string $filename, string $recipient): array
     {
+        deprecationWarning('2.13.4', '`BackupManager::send()` method is deprecated. Will be removed in a future release');
+
         return self::getEmailInstance($filename, $recipient)->send();
     }
 }